+++ src/commonMain/kotlin/com/github/insanusmokrassar/SauceNaoAPI/additional/results/ResultHeader.kt
@@ -1,9 +1,0 @@
-package com.github.insanusmokrassar.SauceNaoAPI.additional.results
-
-import com.github.insanusmokrassar.SauceNaoAPI.additional.header.IndexInfo
-
-data class ResultHeader(
-    val similarity: Float,
-    val thumbnail: String,
-    val index: IndexInfo
-)