+++ src/commonMain/kotlin/com/insanusmokrassar/SauceNaoAPI/models/Result.kt
@@ -1,0 +1,9 @@
+package com.insanusmokrassar.SauceNaoAPI.models
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class Result(
+    val header: ResultHeader,
+    val data: ResultData
+)