+++ src/commonMain/kotlin/com/insanusmokrassar/SauceNaoAPI/models/ResultHeader.kt
@@ -1,0 +1,14 @@
+package com.insanusmokrassar.SauceNaoAPI.models
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ResultHeader(
+    val similarity: Float,
+    val thumbnail: String,
+    @SerialName("index_id")
+    val indexId: Int,
+    @SerialName("index_name")
+    val indexName: String
+)