+++ src/jvmTest/kotlin/Launcher.kt
@@ -1,6 +1,6 @@
-import com.github.insanusmokrassar.SauceNaoAPI.SauceNaoAPI
+import com.insanusmokrassar.SauceNaoAPI.SauceNaoAPI
+import com.insanusmokrassar.SauceNaoAPI.utils.useSafe
 import io.ktor.http.ContentType
-import io.ktor.utils.io.streams.asInput
 import kotlinx.coroutines.*
 import java.io.File
 import java.nio.file.Files
@@ -11,7 +11,7 @@ suspend fun main(vararg args: String) {
     val scope = CoroutineScope(Dispatchers.Default)
 
     val api = SauceNaoAPI(key, scope = scope)
-    api.use { _ ->
+    api.useSafe { _ ->
         println(
             when {
                 requestSubject.startsWith("/") -> File(requestSubject).let {